Clean up Changesets compatibility handling
